Windows port: experimental moe_gpt: fix MSVC narrowing conversion
- moe_gpt_device_operation.cpp: add static_cast<uint32_t> for the size_t
  per_expert_total_tokens_row_elements value placed into a Shape{} braced
  initializer, which MSVC treats as narrowing.

+++ ttnn/cpp/ttnn/operations/experimental/ccl/moe_gpt/device/moe_gpt_device_operation.cpp
@@ -120,7 +120,7 @@ MoEGPTDeviceOperation::spec_return_value_t MoEGPTDeviceOperation::compute_output
     auto per_expert_total_tokens_row_bytes = tt::align(experts_per_device * sizeof(uint32_t), l1_alignment);
     auto per_expert_total_tokens_row_elements = per_expert_total_tokens_row_bytes / sizeof(uint32_t);
     auto per_expert_spec = tt::tt_metal::TensorSpec(
-        Shape({1, per_expert_total_tokens_row_elements}),
+        Shape({1, static_cast<uint32_t>(per_expert_total_tokens_row_elements)}),
         tt::tt_metal::TensorLayout(
             tt::tt_metal::DataType::UINT32,
             tt::tt_metal::PageConfig(tt::tt_metal::Layout::ROW_MAJOR),
